+++ internal/forge/bitbucketcloud/api_connector.go
@@ -1,0 +1,31 @@
+package bitbucketcloud
+
+import (
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+	"github.com/ktrysmt/go-bitbucket"
+)
+
+var _ forgedomain.APIConnector = bbclConnector
+
+func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
+	user, err := self.client.User.Profile()
+	if err != nil {
+		return forgedomain.VerifyConnectionResult{
+			AuthenticatedUser:   None[string](),
+			AuthenticationError: err,
+			AuthorizationError:  nil,
+		}
+	}
+	_, err = self.client.Repositories.PullRequests.Gets(&bitbucket.PullRequestsOptions{
+		Owner:    self.Organization,
+		RepoSlug: self.Repository,
+		Query:    "",
+		States:   []string{},
+	})
+	return forgedomain.VerifyConnectionResult{
+		AuthenticatedUser:   NewOption(user.Username),
+		AuthenticationError: nil,
+		AuthorizationError:  err,
+	}
+}
