+++ internal/forge/bitbucketcloud/proposal_merger.go
@@ -1,0 +1,33 @@
+package bitbucketcloud
+
+import (
+	"errors"
+	"strconv"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	"github.com/ktrysmt/go-bitbucket"
+)
+
+var _ forgedomain.ProposalMerger = bbclConnector
+
+func (self Connector) SquashMergeProposal(number int, message gitdomain.CommitMessage) error {
+	if number <= 0 {
+		return errors.New(messages.ProposalNoNumberGiven)
+	}
+	self.log.Start(messages.ForgeBitbucketMergingViaAPI, colors.BoldGreen().Styled("#"+strconv.Itoa(number)))
+	_, err := self.client.Repositories.PullRequests.Merge(&bitbucket.PullRequestsOptions{
+		ID:       strconv.Itoa(number),
+		Owner:    self.Organization,
+		RepoSlug: self.Repository,
+		Message:  message.String(),
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
