+++ internal/forge/bitbucketcloud/proposal_searcher.go
@@ -1,0 +1,69 @@
+package bitbucketcloud
+
+import (
+	"fmt"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+	"github.com/ktrysmt/go-bitbucket"
+)
+
+var _ forgedomain.ProposalSearcher = bbclConnector
+
+func (self Connector) SearchProposal(branch gitdomain.LocalBranchName) (Option[forgedomain.Proposal], error) {
+	self.log.Start(messages.APIParentBranchLookupStart, branch.String())
+	response1, err := self.client.Repositories.PullRequests.Gets(&bitbucket.PullRequestsOptions{
+		Owner:    self.Organization,
+		RepoSlug: self.Repository,
+		Query:    fmt.Sprintf("source.branch.name = %q", branch),
+		States:   []string{"open"},
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return None[forgedomain.Proposal](), err
+	}
+	response2, ok := response1.(map[string]interface{})
+	if !ok {
+		self.log.Failed(messages.APIUnexpectedResultDataStructure)
+		return None[forgedomain.Proposal](), nil
+	}
+	size1, has := response2["size"]
+	if !has {
+		self.log.Failed(messages.APIUnexpectedResultDataStructure)
+		return None[forgedomain.Proposal](), nil
+	}
+	size2, ok := size1.(float64)
+	if !ok {
+		self.log.Failed(messages.APIUnexpectedResultDataStructure)
+		return None[forgedomain.Proposal](), nil
+	}
+	size3 := int(size2)
+	if size3 == 0 {
+		self.log.Success("none")
+		return None[forgedomain.Proposal](), nil
+	}
+	if size3 > 1 {
+		self.log.Failed(fmt.Sprintf(messages.ProposalMultipleFromFound, size3, branch))
+		return None[forgedomain.Proposal](), nil
+	}
+	values1, has := response2["values"]
+	if !has {
+		self.log.Failed(messages.APIUnexpectedResultDataStructure)
+		return None[forgedomain.Proposal](), nil
+	}
+	values2, ok := values1.([]interface{})
+	if !ok {
+		self.log.Failed(messages.APIUnexpectedResultDataStructure)
+		return None[forgedomain.Proposal](), nil
+	}
+	proposal1 := values2[0].(map[string]interface{})
+	proposal2, err := parsePullRequest(proposal1)
+	if err != nil {
+		self.log.Failed(err.Error())
+		return None[forgedomain.Proposal](), nil
+	}
+	self.log.Success(proposal2.Target.String())
+	return Some(forgedomain.Proposal{Data: proposal2, ForgeType: forgedomain.ForgeTypeBitbucket}), nil
+}
