+++ internal/forge/bitbucketcloud/proposal_updater.go
@@ -1,0 +1,57 @@
+package bitbucketcloud
+
+import (
+	"strconv"
+
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	"github.com/ktrysmt/go-bitbucket"
+)
+
+var _ forgedomain.ProposalUpdater = bbclConnector
+
+func (self Connector) UpdateProposalBody(proposalData forgedomain.ProposalInterface, newBody string) error {
+	data := proposalData.(forgedomain.BitbucketCloudProposalData)
+	self.log.Start(messages.APIProposalUpdateBody, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)))
+	_, err := self.client.Repositories.PullRequests.Update(&bitbucket.PullRequestsOptions{
+		ID:                strconv.Itoa(data.Number),
+		Owner:             self.Organization,
+		RepoSlug:          self.Repository,
+		SourceBranch:      data.Source.String(),
+		DestinationBranch: data.Target.String(),
+		Title:             data.Title,
+		Description:       newBody,
+		Draft:             data.Draft,
+		CloseSourceBranch: data.CloseSourceBranch,
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
+
+func (self Connector) UpdateProposalTarget(proposalData forgedomain.ProposalInterface, target gitdomain.LocalBranchName) error {
+	data := proposalData.(forgedomain.BitbucketCloudProposalData)
+	self.log.Start(messages.APIUpdateProposalTarget, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)), colors.BoldCyan().Styled(target.String()))
+	_, err := self.client.Repositories.PullRequests.Update(&bitbucket.PullRequestsOptions{
+		ID:                strconv.Itoa(data.Number),
+		Owner:             self.Organization,
+		RepoSlug:          self.Repository,
+		SourceBranch:      data.Source.String(),
+		DestinationBranch: target.String(),
+		Title:             data.Title,
+		Description:       data.Body.GetOrZero(),
+		Draft:             data.Draft,
+		CloseSourceBranch: data.CloseSourceBranch,
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
